Fix navigation and login and logout
