+++ src/App.css
@@ -25,10 +25,7 @@ html {
 }
 
 .app__content {
-  background-image: url("./assets/background.jpg");
-  background-position: center;
-  background-size: cover;
-  background-repeat: no-repeat;
+  background-color: rgb(214, 209, 198);
 }
 
 .app__card__container {
@@ -182,4 +179,6 @@ h3.ant-typography {
   color: rgb(77, 55, 9);
   font-weight: 600;
   text-align: center;
+  background-color: pink;
+  /* position: relative; */
 }
