+++ src/Components/Nav.js
@@ -1,63 +1,60 @@
 import React from "react";
-import "../App.css";
-import About from "./About";
-import Products from "./Products";
-import Login from "./Login";
-import Home from "./Home";
-import QuizPage from "../QuizComponents/QuizPage";
-import Logout from "./Logout";
+import { Link } from "react-router-dom";
 
-import { BrowserRouter as Router, Link, Switch, Route } from "react-router-dom";
-import { useSelector } from "react-redux";
-import { selectUser } from "../users/userSlice";
+import "../App.css";
+import { Layout, Menu } from "antd";
+import {
+  QuestionCircleOutlined,
+  SmileOutlined,
+  HomeOutlined,
+  BarcodeOutlined,
+  LoginOutlined,
+  LogoutOutlined,
+} from "@ant-design/icons";
 
 function Nav() {
-  const user = useSelector(selectUser);
-  return (
-    <Router>
-      <div className="nav__container">
-        <div className="nav">
-          <nav>
-            <ul>
-              <li>
-                <Link to="/">Home</Link>
-              </li>
-              <li>
-                <Link to="/about">About</Link>
-              </li>
-              <li>
-                <Link to="/products">Merch</Link>
-              </li>
-              <li>
-                <Link to="/login">Account</Link>
-              </li>
+  const { Header } = Layout;
 
-              <li>
-                <Link to="/quiz">Quiz</Link>
-              </li>
-            </ul>
-          </nav>
-          <Switch>
-            <Route path="/quiz">
-              <QuizPage />
-            </Route>
-            <Route path="/about">
-              <About />
-            </Route>
-            <Route path="/products">
-              <Products />
-            </Route>
-            <Route path="/login">
-              <Login />
-              <div>{user ? <Logout /> : <Login />}</div>
-            </Route>
-            <Route path="/">
-              <Home />
-            </Route>
-          </Switch>
-        </div>
-      </div>
-    </Router>
+  return (
+    <div className="nav__container">
+      <Layout className="layout">
+        <Header>
+          <div className="logo" />
+          <Menu mode="horizontal" defaultSelectedKeys={["2"]}>
+            <Menu.Item key="home">
+              <Link to="/">
+                <HomeOutlined /> Home
+              </Link>
+            </Menu.Item>
+            <Menu.Item key="aboutus">
+              <Link to="/about">
+                <SmileOutlined /> About Us
+              </Link>
+            </Menu.Item>
+            <Menu.Item key="products">
+              <Link to="/products">
+                <BarcodeOutlined /> Merch
+              </Link>
+            </Menu.Item>
+            <Menu.Item key="login">
+              <Link to="/login">
+                <LoginOutlined /> Login
+              </Link>
+            </Menu.Item>
+            <Menu.Item key="logout">
+              <Link to="/logout">
+                <LogoutOutlined /> Logout
+              </Link>
+            </Menu.Item>
+            <Menu.Item key="quiz">
+              <Link to="/quiz">
+                <QuestionCircleOutlined /> Take a Quiz!
+              </Link>
+            </Menu.Item>
+          </Menu>
+        </Header>
+      </Layout>
+    </div>
   );
 }
